move Common.validateArguments to VariantUtils
