+++ src/main/scala/org/hammerlab/guacamole/Common.scala
@@ -24,7 +24,6 @@ import htsjdk.variant.vcf.VCFFileReader
 import org.apache.commons.io.IOUtils
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{FileSystem, Path}
-import org.apache.hadoop.mapred.FileAlreadyExistsException
 import org.apache.spark.{Logging, SparkContext}
 import org.hammerlab.guacamole.distributed.LociPartitionUtils
 import org.hammerlab.guacamole.loci.LociArgs
@@ -209,21 +208,6 @@ object Common extends Logging {
           throw new IllegalArgumentException("Env variables should be key=value syntax, e.g. -spark_env foo=bar")
         }
         a :+ (kvSplit(0), kvSplit(1))
-    }
-  }
-
-  /**
-   * Perform validation of command line arguments at startup.
-   * This allows some late failures (e.g. output file already exists) to be surfaced more quickly.
-   */
-  def validateArguments(args: GenotypeOutputArgs) = {
-    val outputPath = args.variantOutput.stripMargin
-    if (outputPath.toLowerCase.endsWith(".vcf")) {
-      val filesystem = FileSystem.get(new Configuration())
-      val path = new Path(outputPath)
-      if (filesystem.exists(path)) {
-        throw new FileAlreadyExistsException("Output directory " + path + " already exists")
-      }
     }
   }
 }
