+++ src/main/scala/org/hammerlab/guacamole/variants/VariantUtils.scala
@@ -6,6 +6,7 @@ import org.apache.avro.generic.GenericDatumWriter
 import org.apache.avro.io.EncoderFactory
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{FileSystem, Path}
+import org.apache.hadoop.mapred.FileAlreadyExistsException
 import org.apache.spark.rdd.RDD
 import org.bdgenomics.adam.rdd.ADAMContext._
 import org.bdgenomics.formats.avro.{Genotype => BDGGenotype}
@@ -78,5 +79,18 @@ object VariantUtils {
     }
   }
 
-
+  /**
+   * Perform validation of command line arguments at startup.
+   * This allows some late failures (e.g. output file already exists) to be surfaced more quickly.
+   */
+  def validateArguments(args: GenotypeOutputArgs) = {
+    val outputPath = args.variantOutput.stripMargin
+    if (outputPath.toLowerCase.endsWith(".vcf")) {
+      val filesystem = FileSystem.get(new Configuration())
+      val path = new Path(outputPath)
+      if (filesystem.exists(path)) {
+        throw new FileAlreadyExistsException("Output directory " + path + " already exists")
+      }
+    }
+  }
 }
